Reviewed database configuration. Updated Log configuration

+++ Templates/Devon4NetAPI/Devon4Net.Common/Business/TodoManagement/Service/TodoService.cs
@@ -10,37 +10,36 @@ using Devon4Net.Common.Domain.Entities;
 using Devon4Net.Common.Domain.RepositoryInterfaces.TodoManagement;
 using Devon4Net.Domain.UnitOfWork.Service;
 using Devon4Net.Domain.UnitOfWork.UnitOfWork;
+using Devon4Net.Infrastructure.Log;
 using Microsoft.Extensions.Logging;
 
 namespace Devon4Net.Common.Business.TodoManagement.Service
 {
     public class TodoService: Service<TodoContext>, ITodoService
     {
-        private readonly ILogger<TodoService> _logger;
         private readonly ITodoRepository _todoRepository;
 
         public TodoService(IUnitOfWork<TodoContext> uoW, ILogger<TodoService> logger) : base(uoW)
         {
-            _logger = logger;
             _todoRepository = uoW.Repository<ITodoRepository>();
         }
 
         public async Task<IEnumerable<TodoDto>> GetTodo(Expression<Func<Todos, bool>> predicate = null)
         {
-            _logger.LogDebug("GetTodo method from service TodoService");
+            Devon4NetLogger.Debug("GetTodo method from service TodoService");
             var result = await _todoRepository.GetTodo(predicate).ConfigureAwait(false);
             return result.Select(TodoConverter.ModelToDto);
         }
 
         public async Task<Todos> GetTodoById(long id)
         {
-            _logger.LogDebug($"GetTodoById method from service TodoService with value : {id}");
+            Devon4NetLogger.Debug($"GetTodoById method from service TodoService with value : {id}");
             return await _todoRepository.GetTodoById(id).ConfigureAwait(false);
         }
 
         public async Task<Todos> SetTodo(string description)
         {
-            _logger.LogDebug($"SetTodo method from service TodoService with value : {description}");
+            Devon4NetLogger.Debug($"SetTodo method from service TodoService with value : {description}");
 
             if (string.IsNullOrEmpty(description) || string.IsNullOrWhiteSpace(description))
             {
@@ -52,7 +51,14 @@ namespace Devon4Net.Common.Business.TodoManagement.Service
 
         public async Task<long> DeleteTodoById(long id)
         {
-            _logger.LogDebug($"DeleteTodoById method from service TodoService with value : {id}");
+            Devon4NetLogger.Debug($"DeleteTodoById method from service TodoService with value : {id}");
+            var todo = await _todoRepository.GetFirstOrDefault(t => t.Id == id).ConfigureAwait(false);
+
+            if (todo == null)
+            {
+                throw new ArgumentException($"The provided Id {id} does not exists");
+            }
+
             return await _todoRepository.DeleteTodoById(id).ConfigureAwait(false);
         }
     }
